requirements.txt
Flask==1.1.2
waitress==2.0.0
requests==2.25.1
numpy==1.19.5
matplotlib==3.3.3
pytesseract==0.3.7
Pillow==8.2.0
opencv-python-headless
#tqdm>=4.41.0
#PyYAML>=5.3.1
#scipy>=1.4.1
#torchvision==0.8.1
#tensorboard>=2.4.1
#seaborn>=0.11.0
openvino==2021.3.0
numba
#easyocr==1.3.1
#thop
